Add files via upload
This study looks at a company to determine how it is doing right now in terms of pay equity and gender pay. The purpose of this report is to provide an answer to the commercial query, "Is there a wage disparity in the company?" Other queries include, "What is the situation regarding gender pay and equal pay in the company," and "What recommendations can be given to address such issues in the pay gap to ensure equal pay in compliance with the UK legal requirements."
The Equal Pay Act defines work as having an equal value regardless of the person conducting the task, according to this definition.
According to UK law, the gender pay gap is the difference between the average hourly wages earned by men and women workers across the economy. "Salary" is the primary performance indicator used in this report. This variable was chosen as the outcome variable because it provides the most comprehensive explanation of whether there is a wage gap between men and women inside the organization. Salary is used as a performance indicator to highlight pay disparities.


The analysis's findings allowed the report to demonstrate that the outcome variable The gender pay gap and pay discrepancy in the organization are problems, but salary has no bearing on them. In light of the fact that there are additional factors influencing the gender pay gap, the company should review its payroll system, ensure that a more defined and transparent system is put in place to monitor employee wages, and consider hiring more female employees.